refactor: Separate quiz form and list logic
- Updated `Quizzes.jsx` component to toggle form visibility and handle quiz editing.
- Added functionality to `QuizForm.jsx` for creating and updating quizzes.
- Passed quiz data to `QuizForm.jsx` for editing purposes.
- Implemented `onClose` callback to manage form visibility from `QuizForm.jsx`.
